chore: nextjs cleanup — inline demo, trim scaffold-eth leftovers
- Inline app/_components/FHECounterDemo.tsx into app/page.tsx and
  remove the _components/ directory. The component was only used from
  page.tsx so the separate file added indirection for no reason.
- Delete unused util modules: utils/helper/block.ts, contractsData.ts,
  decodeTxData.ts, getParsedError.ts, notification.tsx. Zero references
  from app code after the hardhat → foundry / SDK v2 refactor.
- Rewrite utils/helper/contract.ts (395 → 47 lines). Keeps only what
  useFHECounterWagmi and useDeployedContractInfo actually import
  (Contract, ContractName, ContractCodeStatus, UseDeployedContractConfig,
  contracts, GenericContract, GenericContractsDeclaration,
  InheritedFunctions). Drops all scaffold-eth boilerplate for hooks the
  template doesn't use: UseScaffoldReadConfig, UseScaffoldWriteConfig,
  UseScaffoldEventConfig, UseScaffoldEventHistoryConfig,
  FunctionNamesWithInputs, ScaffoldWriteContractOptions,
  getParsedErrorWithAllAbis, simulateContractWriteAndNotifyError, etc.
- Update utils/helper/index.ts barrel to only re-export networks (the
  other modules it re-exported are now gone).
- Drop packages/nextjs/vercel.json — specified "installCommand":
  "yarn install" on a pnpm project. Vercel auto-detects from
  pnpm-lock.yaml.
- Drop type-fest devDep (only used by the removed scaffold types).

Verified: pnpm next:check-types and pnpm next:build both green.
