+++ packages/nextjs/package.json
@@ -49,7 +49,6 @@
     "postcss": "~8.4.45",
     "prettier": "~3.5.3",
     "tailwindcss": "4.1.3",
-    "type-fest": "~4.26.1",
     "typescript": "~5.8.2",
     "vercel": "~39.1.3"
   }
