--- packages/nextjs/utils/helper/block.ts
@@ -1,17 +1,0 @@
-import { Block, Transaction, TransactionReceipt } from "viem";
-
-export type TransactionWithFunction = Transaction & {
-  functionName?: string;
-  functionArgs?: any[];
-  functionArgNames?: string[];
-  functionArgTypes?: string[];
-};
-
-type TransactionReceipts = {
-  [key: string]: TransactionReceipt;
-};
-
-export type TransactionsTableProps = {
-  blocks: Block[];
-  transactionReceipts: TransactionReceipts;
-};
