--- packages/nextjs/utils/helper/contractsData.ts
@@ -1,11 +1,0 @@
-import { useTargetNetwork } from "~~/hooks/helper";
-import { GenericContractsDeclaration, contracts } from "~~/utils/helper/contract";
-
-const DEFAULT_ALL_CONTRACTS: GenericContractsDeclaration[number] = {};
-
-export function useAllContracts() {
-  const { targetNetwork } = useTargetNetwork();
-  const contractsData = contracts?.[targetNetwork.id];
-  // using constant to avoid creating a new object on every call
-  return contractsData || DEFAULT_ALL_CONTRACTS;
-}
