--- packages/nextjs/vercel.json
@@ -1,3 +1,0 @@
-{
-  "installCommand": "yarn install"
-}
